// src/forum/topics/interfaces/topic-response.interface.ts
import { IAuditFields } from "src/common/interfaces/common.interface";

export interface ITopicResponse extends IAuditFields {
    id: string;
    title: string;
    slug: string;
    content: string;
    tagId: string;
    tagTitle: string;
    tagSlug: string;
    tagDescription: string;
    userId: string;
    author: string;
    authorUsername: string;
    authorNickname: string;
    authorBio: string;
    authorRole: string;

    viewCount: number;

    lastPostAt: Date;
    postCount: number;
    status: boolean;
}